Folder/file select and organized output saving
Original system of processing for osa and for other. Now allows user to choose to run an individual file or run through a folder. If user selects folder, user can then choose to run OSA files, LIV/WLM files, or both. OSA files are identified by "osa" in the file name or if they are larger than 50KB. Files which are not OSA are assumed to be LIV or WLM.