Added Subfolder search and error handling
System will look through selected folder and if a file is located in a subfolder it will save all outputs to the same subfolder. If the file is not in a subfolder it will create a subfolder with the file's name and save outputs there.

Furthermore, if when traversing a folder a processing error comes up for one of the files, the file name and a summary of the error is printed and the file is skipped. System continues traversing the folder.